[CCP-153] moved domain classes to old package structure
