+++ Conceptpower+Spring/src/main/java/edu/asu/conceptpower/root/TypeDatabaseClient.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Component;
 import com.db4o.ObjectContainer;
 import com.db4o.ObjectSet;
 
-import edu.asu.conceptpower.servlet.core.ConceptType;
+import edu.asu.conceptpower.core.ConceptType;
 
 @Component
 public class TypeDatabaseClient {
