+++ Conceptpower+Spring/src/main/java/edu/asu/conceptpower/servlet/bean/ConceptEditBean.java
@@ -6,7 +6,7 @@ import java.util.Map;
 
 import org.springframework.stereotype.Component;
 
-import edu.asu.conceptpower.servlet.core.ConceptType;
+import edu.asu.conceptpower.core.ConceptType;
 import edu.asu.conceptpower.servlet.core.POS;
 
 @Component
