+++ Conceptpower+Spring/src/main/java/edu/asu/conceptpower/servlet/core/impl/ConceptListManager.java
@@ -6,12 +6,10 @@ import java.util.UUID;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import edu.asu.conceptpower.servlet.core.ConceptList;
+import edu.asu.conceptpower.core.ConceptList;
 import edu.asu.conceptpower.servlet.core.IConceptListManager;
 import edu.asu.conceptpower.servlet.db4o.DBNames;
 import edu.asu.conceptpower.servlet.db4o.IConceptDBManager;
-import edu.asu.conceptpower.servlet.exceptions.DictionaryExistsException;
-import edu.asu.conceptpower.servlet.wordnet.Constants;
 
 /**
  * Manager class for concept lists.
