+++ Conceptpower+Spring/src/main/java/edu/asu/conceptpower/servlet/db/objectdb/ConceptDBManager.java
@@ -4,8 +4,8 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 
-import edu.asu.conceptpower.servlet.core.ConceptEntry;
-import edu.asu.conceptpower.servlet.core.ConceptList;
+import edu.asu.conceptpower.core.ConceptEntry;
+import edu.asu.conceptpower.core.ConceptList;
 import edu.asu.conceptpower.servlet.db.IDatabaseManager;
 import edu.asu.conceptpower.servlet.db4o.IConceptDBManager;
