+++ Conceptpower+Spring/src/main/java/edu/asu/conceptpower/servlet/db4o/IConceptDBManager.java
@@ -2,8 +2,8 @@ package edu.asu.conceptpower.servlet.db4o;
 
 import java.util.List;
 
-import edu.asu.conceptpower.servlet.core.ConceptEntry;
-import edu.asu.conceptpower.servlet.core.ConceptList;
+import edu.asu.conceptpower.core.ConceptEntry;
+import edu.asu.conceptpower.core.ConceptList;
 
 public interface IConceptDBManager {
