+++ Conceptpower+Spring/src/main/java/edu/asu/conceptpower/servlet/rest/Concepts.java
@@ -18,7 +18,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
-import edu.asu.conceptpower.servlet.core.ConceptEntry;
+import edu.asu.conceptpower.core.ConceptEntry;
 import edu.asu.conceptpower.servlet.core.IConceptManager;
 import edu.asu.conceptpower.servlet.core.POS;
 import edu.asu.conceptpower.servlet.exceptions.DictionaryDoesNotExistException;
