+++ Conceptpower+Spring/src/main/java/edu/asu/conceptpower/servlet/rest/TypeIdLookup.java
@@ -12,8 +12,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
 
+import edu.asu.conceptpower.core.ConceptType;
 import edu.asu.conceptpower.root.TypeDatabaseClient;
-import edu.asu.conceptpower.servlet.core.ConceptType;
 import edu.asu.conceptpower.servlet.xml.XMLConstants;
 import edu.asu.conceptpower.servlet.xml.XMLMessageFactory;
 import edu.asu.conceptpower.servlet.xml.XMLTypeMessage;
