+++ Conceptpower+Spring/src/main/java/edu/asu/conceptpower/servlet/web/ConceptTypeEditController.java
@@ -19,7 +19,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
-import edu.asu.conceptpower.servlet.core.ConceptType;
+import edu.asu.conceptpower.core.ConceptType;
 import edu.asu.conceptpower.servlet.core.IConceptTypeManger;
 import edu.asu.conceptpower.servlet.users.IUserManager;
 import edu.asu.conceptpower.servlet.validation.ConceptTypeAddValidator;
