+++ Conceptpower+Spring/src/main/java/edu/asu/conceptpower/servlet/wrapper/ConceptTypeWrapper.java
@@ -4,8 +4,8 @@ import java.io.Serializable;
 
 import org.springframework.beans.factory.annotation.Autowired;
 
+import edu.asu.conceptpower.core.ConceptType;
 import edu.asu.conceptpower.root.URIHelper;
-import edu.asu.conceptpower.servlet.core.ConceptType;
 
 /**
  *
