+++ Conceptpower+Spring/src/main/java/edu/asu/conceptpower/servlet/wrapper/impl/ConceptEntryWrapperCreator.java
@@ -6,8 +6,8 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
+import edu.asu.conceptpower.core.ConceptEntry;
 import edu.asu.conceptpower.root.IURIHelper;
-import edu.asu.conceptpower.servlet.core.ConceptEntry;
 import edu.asu.conceptpower.servlet.core.Constants;
 import edu.asu.conceptpower.servlet.core.IConceptManager;
 import edu.asu.conceptpower.servlet.core.IConceptTypeManger;
