+++ Conceptpower+Spring/src/test/java/edu/asu/conceptpower/core/impl/ConceptTypesManagerTest.java
@@ -11,8 +11,8 @@ import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
 
+import edu.asu.conceptpower.core.ConceptType;
 import edu.asu.conceptpower.root.TypeDatabaseClient;
-import edu.asu.conceptpower.servlet.core.ConceptType;
 import edu.asu.conceptpower.servlet.core.impl.ConceptTypesManager;
 
 public class ConceptTypesManagerTest {
